+++ app/Admin/Controllers/Subway/CityController.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace App\Admin\Controllers;
+namespace App\Admin\Controllers\Subway;
 
-use App\Models\Tag;
-use App\Models\Video;
+use App\Models\Subway\City;
 use Encore\Admin\Form;
 use Encore\Admin\Grid;
 use Encore\Admin\Facades\Admin;
 use Encore\Admin\Layout\Content;
 use App\Http\Controllers\Controller;
 use Encore\Admin\Controllers\ModelForm;
+use Encore\Admin\Show;
 
-class VideoController extends Controller
+class CityController extends Controller
 {
     use ModelForm;
 
@@ -23,18 +23,43 @@ class VideoController extends Controller
     public function index()
     {
         return Admin::content(function (Content $content) {
-            $content->header('header');
+
+            $content->header('Cities');
             $content->description('description');
-            $content->body($this->grid());
+
+            $content->row(function($row) {
+                $row->column(10, $this->grid());
+                $row->column(2, view('admin.grid.subway'));
+            });
         });
     }
 
     public function show($id)
     {
         return Admin::content(function (Content $content) use ($id) {
-            $content->header('header');
-            $content->description('description');
-            $content->body(Admin::show(Video::findOrFail($id)));
+
+            $content->header('Cities');
+            $content->description('城市详情');
+
+            $content->body(Admin::show(City::findOrFail($id), function (Show $show) {
+
+                $show->cn_name('中文名');
+                $show->en_name('英文名');
+                $show->code('城市编码');
+                $show->field('pre');
+
+                $show->lines('地铁线', function ($line) {
+
+                    $line->resource('/demo/subway/lines');
+
+                    $line->id();
+                    $line->name();
+                    $line->pair_uid();
+                    $line->created_at();
+                    $line->updated_at();
+
+                });
+            }));
         });
     }
 
@@ -78,23 +103,27 @@ class VideoController extends Controller
      */
     protected function grid()
     {
-        return Admin::grid(Video::class, function (Grid $grid) {
+        return Admin::grid(City::class, function (Grid $grid) {
 
             $grid->id('ID')->sortable();
 
-            $grid->title()->limit(30);
+            $grid->cn_name();
+            $grid->en_name();
 
-            $grid->status()->radio([
-                0 => 'Sed ut perspiciatis unde omni',
-                1 => 'voluptatem accusantium doloremque',
-                2 => 'dicta sunt explicabo',
-                3 => 'laudantium, totam rem aperiam',
-            ]);
+//            $grid->lines()->display(function ($lines) {
+//                return array_column($lines, 'name');
+//            })->implode('</br>');
 
-            $grid->tags()->pluck('name')->label();
+//            $grid->code();
+//            $grid->pre();
 
             $grid->created_at();
             $grid->updated_at();
+
+            $grid->filter(function ($filter) {
+                $filter->expand();
+                $filter->like('cn_name', 'Name');
+            });
         });
     }
 
@@ -105,24 +134,9 @@ class VideoController extends Controller
      */
     protected function form()
     {
-        return Admin::form(Video::class, function (Form $form) {
+        return Admin::form(City::class, function (Form $form) {
 
             $form->display('id', 'ID');
-
-            $form->text('title')->rules('required');
-
-            $form->radio('status')->options([
-                0 => 'Sed ut perspiciatis unde omni',
-                1 => 'voluptatem accusantium doloremque',
-                2 => 'dicta sunt explicabo',
-                3 => 'laudantium, totam rem aperiam',
-            ])->stacked();
-
-            $form->file('video');
-
-            $form->datetime('release_at');
-
-            $form->multipleSelect('tags')->options(Tag::all()->pluck('name', 'id'));
 
             $form->display('created_at', 'Created At');
             $form->display('updated_at', 'Updated At');