+++ app/Admin/Extensions/Nav/Links.php
@@ -7,7 +7,6 @@ class Links
     public function __toString()
     {
         return <<<HTML
-
 <li>
     <a href="#">
       <i class="fa fa-envelope-o"></i>