+++ app/Models/PostComment.php
@@ -10,6 +10,11 @@ class PostComment extends Model
 
     protected $fillable = ['content'];
 
+    public function scopeOfPost($query, $post)
+    {
+        return $query->where('post_id', $post);
+    }
+
     public function post()
     {
         return $this->belongsTo(Post::class, 'post_id');