+++ app/Models/User/Address.php
@@ -2,6 +2,7 @@
 
 namespace App\Models\User;
 
+use App\Models\ChinaArea;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Model;
 
@@ -12,5 +13,20 @@ class Address extends Model
     public function user()
     {
         return $this->belongsTo(User::class);
+    }
+
+    public function province()
+    {
+        return $this->belongsTo(ChinaArea::class);
+    }
+
+    public function city()
+    {
+        return $this->belongsTo(ChinaArea::class);
+    }
+
+    public function district()
+    {
+        return $this->belongsTo(ChinaArea::class);
     }
 }