+++ public/js/redpacket.js
@@ -1,0 +1,4 @@
+var img = document.createElement("img");
+img.src = '/img/redpacket.png';
+img.style.cssText = "bottom:5px;right:5px;position:fixed;max-height:250px;";
+document.body.appendChild(img);