+++ resources/views/admin/grid/card.blade.php
@@ -1,21 +1,24 @@
 <div class="box">
-    <div class="box-header">
+    @if(isset($title))
+        <div class="box-header with-border">
+            <h3 class="box-title"> {{ $title }}</h3>
+        </div>
+    @endif
 
-        <h3 class="box-title"></h3>
-
-        <span style="position: absolute;left: 10px;top: 5px;">
-            {!! $grid->renderHeaderTools() !!}
-        </span>
-
-        <div class="box-tools">
-            {!! $grid->renderFilter() !!}
+    <div class="box-header with-border">
+        <div class="pull-right">
             {!! $grid->renderExportButton() !!}
             {!! $grid->renderCreateButton() !!}
         </div>
-
+        <span>
+            {!! $grid->renderHeaderTools() !!}
+        </span>
     </div>
+
+    {!! $grid->renderFilter() !!}
+
     <!-- /.box-header -->
-    <div class="box-footer">
+    <div class="box-body">
         <ul class="mailbox-attachments clearfix">
             @foreach($grid->rows() as $row)
                 <li>