+++ public/docs/en/change-log.md
@@ -1,4 +1,15 @@
-# Change log
+# Release Notes
+
+## v1.5.18 (2018-08-10)
+
+- Refactored the style of the grid filter, adding `scope` query support
+- `Model-show` supports displaying fields as file style
+- Update editable empty text style
+- Support for displaying two-dimensional array fields in `Model-grid` as table
+
+## v1.5.16 (2018-08-3)
+
+- Add `Model-show` to support display data details
 
 ## v1.2.9、v1.3.3、v1.4.1